Core/Entity A large cherry-pick form the TrinityCore master

+++ src/server/game/Server/Packets/AchievementPackets.cpp
@@ -33,8 +33,8 @@ ByteBuffer& operator<<(ByteBuffer& data, WorldPackets::Achievement::CriteriaProg
     data << uint64(criteria.Quantity);
     data << criteria.Player;
     data.AppendPackedTime(criteria.Date);
-    data << uint32(criteria.TimeFromStart);
-    data << uint32(criteria.TimeFromCreate);
+    data << criteria.TimeFromStart;
+    data << criteria.TimeFromCreate;
     data.WriteBits(criteria.Flags, 4);
     data.FlushBits();
     return data;
@@ -76,8 +76,8 @@ WorldPacket const* WorldPackets::Achievement::CriteriaUpdate::Write()
     _worldPacket << PlayerGUID;
     _worldPacket << uint32(Flags);
     _worldPacket.AppendPackedTime(CurrentTime);
-    _worldPacket << uint32(ElapsedTime);
-    _worldPacket << uint32(CreationTime);
+    _worldPacket << ElapsedTime;
+    _worldPacket << CreationTime;
 
     return &_worldPacket;
 }
@@ -129,8 +129,8 @@ WorldPacket const* WorldPackets::Achievement::GuildCriteriaUpdate::Write()
     for (GuildCriteriaProgress const& progress : Progress)
     {
         _worldPacket << int32(progress.CriteriaID);
-        _worldPacket << uint32(progress.DateCreated);
-        _worldPacket << uint32(progress.DateStarted);
+        _worldPacket << progress.DateCreated;
+        _worldPacket << progress.DateStarted;
         _worldPacket.AppendPackedTime(progress.DateUpdated);
         _worldPacket << uint64(progress.Quantity);
         _worldPacket << progress.PlayerGUID;
