+++ src/tools/vmap4_extractor/adtfile.h
@@ -1,6 +1,5 @@
 /*
- * Copyright (C) 2008-2018 TrinityCore <https://www.trinitycore.org/>
- * Copyright (C) 2005-2011 MaNGOS <http://getmangos.com/>
+ * This file is part of the TrinityCore Project. See AUTHORS file for Copyright information
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -53,7 +52,7 @@ namespace ADT
 
 struct ADTOutputCache
 {
-    uint32 Flags;
+    uint8 Flags;
     std::vector<uint8> Data;
 };
